+++ src/ridgeplot/_color/colorscale.py
@@ -1,0 +1,52 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any, cast
+
+import plotly.express as px
+from _plotly_utils.basevalidators import ColorscaleValidator as _ColorscaleValidator
+
+from ridgeplot._types import Color, ColorScale
+
+if TYPE_CHECKING:
+    from collections.abc import Collection
+
+
+def list_all_colorscale_names() -> list[str]:
+    """Get a list with all available colorscale names.
+
+    .. versionadded:: 0.1.21
+        Replaced the deprecated function ``get_all_colorscale_names()``.
+
+    Returns
+    -------
+    list[str]
+        A list with all available colorscale names.
+    """
+    # Add 'default' for backwards compatibility
+    px_colorscales = px.colors.named_colorscales()
+    return sorted({"default", *px_colorscales, *(f"{name}_r" for name in px_colorscales)})
+
+
+class ColorscaleValidator(_ColorscaleValidator):  # type: ignore[misc]
+    def __init__(self) -> None:
+        super().__init__("colorscale", "ridgeplot")
+
+    @property
+    def named_colorscales(self) -> dict[str, list[str]]:
+        named_colorscales = cast(dict[str, list[str]], super().named_colorscales)
+        if "default" not in named_colorscales:
+            # Add 'default' for backwards compatibility
+            named_colorscales["default"] = px.colors.DEFAULT_PLOTLY_COLORS
+        return named_colorscales
+
+    def validate_coerce(self, v: Any) -> ColorScale:
+        coerced = super().validate_coerce(v)
+        if coerced is None:
+            self.raise_invalid_val(coerced)
+        return cast(ColorScale, [tuple(c) for c in coerced])
+
+
+def validate_and_coerce_colorscale(colorscale: ColorScale | Collection[Color] | str) -> ColorScale:
+    """Convert mixed colorscale representations to the canonical
+    :data:`ColorScale` format."""
+    return ColorscaleValidator().validate_coerce(colorscale)
