+++ src/ridgeplot/_color/css_colors.py
