+++ tests/unit/color/test_utils.py
@@ -5,19 +5,11 @@ from typing import TYPE_CHECKING, Any
 import pytest
 from _plotly_utils.exceptions import PlotlyError
 
-from ridgeplot._coloring.colors import (
-    Color,
-    ColorScale,
-    apply_alpha,
-    interpolate_color,
-    list_all_colorscale_names,
-    round_color,
-    to_rgb,
-    validate_and_coerce_colorscale,
-)
+from ridgeplot._color.utils import apply_alpha, round_color, to_rgb
 
 if TYPE_CHECKING:
-    from collections.abc import Collection
+
+    from ridgeplot._types import Color
 
 
 # ==============================================================
@@ -81,67 +73,6 @@ def test_to_rgb_bug_in_validation_incomplete(
 ) -> None:
     with pytest.raises(expected_exception, match=exception_match or ""):
         to_rgb(color=color)
-
-
-# ==============================================================
-# --- list_all_colorscale_names()
-# ==============================================================
-
-
-def test_list_all_colorscale_names() -> None:
-    all_colorscale_names = list_all_colorscale_names()
-    assert all(isinstance(name, str) for name in all_colorscale_names)
-    assert "viridis" in all_colorscale_names
-    assert "default" in all_colorscale_names
-    for name in all_colorscale_names:
-        validate_and_coerce_colorscale(name)
-
-
-# ==============================================================
-# ---  validate_and_coerce_colorscale()
-# ==============================================================
-
-
-def test_validate_and_coerce_colorscale(
-    valid_colorscale: tuple[ColorScale | Collection[Color] | str, ColorScale]
-) -> None:
-    colorscale, expected = valid_colorscale
-    coerced = validate_and_coerce_colorscale(colorscale=colorscale)
-    values, colors = zip(*coerced)
-    values_expected, colors_expected = zip(*expected)
-    assert values == pytest.approx(values_expected)
-    assert colors == colors_expected
-
-
-def test_validate_and_coerce_colorscale_fails(
-    invalid_colorscale: ColorScale | Collection[Color] | str,
-) -> None:
-    with pytest.raises(
-        ValueError, match=r"Invalid value .* received for the 'colorscale' property"
-    ):
-        validate_and_coerce_colorscale(invalid_colorscale)
-
-
-# ==============================================================
-# ---  interpolate_color()
-# ==============================================================
-
-
-def test_interpolate_color_p_in_scale(viridis_colorscale: ColorScale) -> None:
-    viridis_colorscale = list(viridis_colorscale)
-    assert interpolate_color(colorscale=viridis_colorscale, p=0) == viridis_colorscale[0][1]
-    assert interpolate_color(colorscale=viridis_colorscale, p=1) == viridis_colorscale[-1][1]
-
-
-def test_interpolate_color_p_not_in_scale(viridis_colorscale: ColorScale) -> None:
-    # Hard-coded test case.
-    assert interpolate_color(colorscale=viridis_colorscale, p=0.5) == "rgb(34.5, 144.0, 139.5)"
-
-
-@pytest.mark.parametrize("p", [-10.0, -1.3, 1.9, 100.0])
-def test_interpolate_color_fails_for_p_out_of_bounds(p: float) -> None:
-    with pytest.raises(ValueError, match="should be a float value between 0 and 1"):
-        interpolate_color(colorscale=..., p=p)  # type: ignore[arg-type]
 
 
 # ==============================================================
